chore: Move configloader to runtime
